extract heightmap window and gltexture classes
