+++ gui-imgui/src/heightmapPreview.h
@@ -1,0 +1,10 @@
+#pragma once
+#include <GLFW/glfw3.h>
+
+#include "glTexture.h"
+#include "guiState.h"
+
+class HeightmapPreviewWindow {
+public:
+    void render(const GuiState& state, const GLTexture& tex);
+};
