+++ gui-imgui/src/main.cpp
@@ -7,40 +7,15 @@
 #include "backends/imgui_impl_opengl3.h"
 #include "flat/flatTerrainGenerator.h"
 #include "generatorFactory.h"
+#include "glTexture.h"
 #include "guiState.h"
+#include "heightmapPreview.h"
 #include "perlin/perlinTerrainGenerator.h"
 #include "terrainControls.h"
 
 #include <GLFW/glfw3.h>
 #include <vector>
-#include <cstdint>
 #include <iostream>
-
-
-
-// --------------------------
-// OpenGL texture helper
-// --------------------------
-class GLTexture {
-public:
-    GLuint id = 0;
-
-    ~GLTexture() {
-        if (id) glDeleteTextures(1, &id);
-    }
-
-    void upload(const std::vector<float>& pixels, int width, int height) {
-        if (id) glDeleteTextures(1, &id);
-        glGenTextures(1, &id);
-        glBindTexture(GL_TEXTURE_2D, id);
-        glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MIN_FILTER, GL_LINEAR);
-        glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MAG_FILTER, GL_LINEAR);
-        glTexImage2D(GL_TEXTURE_2D, 0, GL_RGB, width, height, 0, GL_RGB, GL_FLOAT, pixels.data());
-        glBindTexture(GL_TEXTURE_2D, 0);
-    }
-
-    operator bool() const { return id != 0; }
-};
 
 // --------------------------
 // Utility to flatten Heightmap
@@ -61,39 +36,6 @@ std::vector<float> flattenHeightmap(const Heightmap& hm) {
 }
 
 
-// --------------------------
-// Heightmap preview window
-// --------------------------
-class HeightmapPreviewWindow {
-public:
-    void render(const GuiState& state, const GLTexture& tex) {
-        if (!tex) return;
-
-        ImGui::Begin("Heightmap Preview");
-
-        // Get available content region size (space inside the window)
-        ImVec2 availSize = ImGui::GetContentRegionAvail();
-
-        // Compute aspect ratio based on terrain grid
-        float aspect = static_cast<float>(state.gridSize.x) / static_cast<float>(state.gridSize.y);
-
-        ImVec2 imageSize;
-        if (availSize.x / availSize.y > aspect) {
-            // Window is wider than terrain; fit to height
-            imageSize.y = availSize.y;
-            imageSize.x = imageSize.y * aspect;
-        } else {
-            // Window is taller than terrain; fit to width
-            imageSize.x = availSize.x;
-            imageSize.y = imageSize.x / aspect;
-        }
-
-        ImGui::Text("Generated Heightmap:");
-        ImGui::Image((void*)(intptr_t)tex.id, imageSize);
-
-        ImGui::End();
-    }
-};
 
 // --------------------------
 // Main
